rename: useTabFromUrl hook을 전역으로 쓰기 위해 폴더 이동 src/features/media/hooks => src/assets/hooks
