+++ src/pages/media/News.tsx
@@ -4,8 +4,8 @@ import SearchBar from '@/features/media/common/SearchBar';
 import NewsContent from '@/features/media/components/news/NewsContent';
 import PressContent from '@/features/media/components/news/PressContent';
 
+import { useTabFromUrl } from '@/assets/hooks/useTabFromUrl';
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui';
-import { useTabFromUrl } from '@/features/media/hooks/useTabFromUrl';
 import { cn } from '@/lib/utils';
 
 import '@/features/media/css/media.css';
